separate fillRow.js. plotSettings to config.
