--- src/chart/AETable/eventListeners.js
@@ -1,3 +1,0 @@
-export function eventListeners(data, vars, settings) {
-
-}
