+++ src/chart/util.js
@@ -6,9 +6,11 @@ import { calculateDifference } from './util/calculateDifference';
 import { addDifferences } from './util/addDifferences';
 import { cross } from './util/cross';
 import { sort } from './util/sort';
+import { fillRow } from './util/fillRow'
 
 export const util =
     {calculateDifference: calculateDifference
     ,addDifferences: addDifferences
     ,cross: cross
-    ,sort: sort};
+    ,sort: sort
+	,fillRow: fillRow};
